+++ requirements.txt
@@ -18,3 +18,4 @@ tensorflow==2.8.*
 tqdm==4.64.*
 protobuf==3.19.*
 PyYAML==6.*
+Jinja2
