+++ src/syngen/ml/metrics/__init__.py
@@ -1,6 +1,6 @@
 from syngen.ml.metrics.metrics_classes.metrics import BaseMetric
-from syngen.ml.metrics.metrics_classes.metrics import BivariateMetric
-from syngen.ml.metrics.metrics_classes.metrics import UnivariateMetric
+from syngen.ml.metrics.metrics_classes.bivariate_metric import BivariateMetric
+from syngen.ml.metrics.metrics_classes.univariate_metric import UnivariateMetric
 from syngen.ml.metrics.metrics_classes.metrics import JensenShannonDistance
 from syngen.ml.metrics.metrics_classes.metrics import Correlations
 from syngen.ml.metrics.accuracy_test.accuracy_test import BaseTest
